:refactor: moved arima function into basepredictor template
