+++ btc_predictor/btc_predictor/models/lstm.py
@@ -233,7 +233,6 @@ class LSTMBTCPredictor(BasePredictor):
         Returns:
             np.ndarray: log return numpy array.
         """
-        _logger.info(f"{type(df)}")
         df["log_ret"] = np.log(df["close"]) - np.log(df["close"].shift(1))
         df.dropna(inplace=True)
         return np.diff(df["log_ret"].to_numpy()).astype("float32")
